+++ react/typings/graphql.d.ts
@@ -1,4 +1,0 @@
-declare module '*.graphql' {
-  const value: any
-  export default value
-}