+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [2.146.0] - 2025-12-18
+## [2.146.0] - 2025-12-18 [YANKED]
 
 ### Added